moved logs into package
