+++ BSS/src/log/DepositLog.java
@@ -1,22 +1,19 @@
-package bss;
+package log;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.time.LocalDateTime;
 
-public class AccountCreationLog implements Log {
+public class DepositLog implements Log {
 	
 	private int id;
-	private int teller_ID;
     private String message;
     private LocalDateTime timeStamp;
     private int accountID;
-    
-    public AccountCreationLog(int id, int teller_ID, String message, 
-    		LocalDateTime timeStamp, int accountID) {
-    	
-        this.id = id;
-        this.teller_ID = teller_ID;
+
+    public DepositLog(int id, String message, LocalDateTime timeStamp, int accountID) {
+        
+    	this.id = id;
         this.message = message;
         this.timeStamp = timeStamp;
         this.accountID = accountID;
@@ -28,16 +25,16 @@ public class AccountCreationLog implements Log {
         try (FileWriter writer = new FileWriter(inOutFile, true)) 
         {
             writer.write("Log ID: " + id + "\n");
-            writer.write("Teller ID: " + teller_ID + "\n");
             writer.write("Message: " + message + "\n");
             writer.write("Timestamp: " + timeStamp.toString() + "\n");
             writer.write("Account ID: " + accountID + "\n");
             writer.write("----\n");
+            
         } 
         catch (IOException e) 
         {
             e.printStackTrace();
         }
     }
-
+    
 }
