+++ BSS/src/log/Log.java
@@ -1,4 +1,4 @@
-package bss;
+package log;
 import java.util.Scanner;
 import java.io.FileWriter;
 import java.io.File;
