+++ BSS/src/log/SessionLog.java
@@ -1,38 +1,41 @@
-package bss;
+package log;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.time.LocalDateTime;
 
-public class WithdrawLog implements Log {
-
+public class SessionLog implements Log {
+   
 	private int id;
     private String message;
-    private LocalDateTime timeStamp;
+    private LocalDateTime timeStart;
+    private LocalDateTime timeEnd;
     private int accountID;
 
-    public WithdrawLog(int id, String message, LocalDateTime timeStamp, int accountID) {
+    public SessionLog(int id, String message, 
+    		LocalDateTime timeStart, LocalDateTime timeEnd, int accountID) {
         
     	this.id = id;
         this.message = message;
-        this.timeStamp = timeStamp;
+        this.timeStart = timeStart;
+        this.timeEnd = timeEnd;
         this.accountID = accountID;
     }
     
     @Override
     public void writeLogToFile(File inOutFile) 
     {
-        try (FileWriter writer = new FileWriter(inOutFile, true)) 
-        {
+    	try (FileWriter writer = new FileWriter(inOutFile, true)) 
+    	{
             writer.write("Log ID: " + id + "\n");
             writer.write("Message: " + message + "\n");
-            writer.write("Timestamp: " + timeStamp.toString() + "\n");
+            writer.write("Session Start: " + timeStart.toString() + "\n");
+            writer.write("Session End: " + timeEnd.toString() + "\n");
             writer.write("Account ID: " + accountID + "\n");
             writer.write("----\n");
-            
         } 
-        catch (IOException e) 
-        {
+    	catch (IOException e) 
+    	{
             e.printStackTrace();
         }
     }
