+++ BSS/src/log/TransferLog.java
@@ -1,22 +1,25 @@
-package bss;
+package log;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.time.LocalDateTime;
 
-public class DepositLog implements Log {
+public class TransferLog implements Log {
 	
 	private int id;
     private String message;
     private LocalDateTime timeStamp;
     private int accountID;
-
-    public DepositLog(int id, String message, LocalDateTime timeStamp, int accountID) {
-        
-    	this.id = id;
+    private int toAccountID;
+    
+    public TransferLog(int id, String message, LocalDateTime timeStamp, 
+    		int accountID, int toAccountID) {
+    	
+        this.id = id;
         this.message = message;
         this.timeStamp = timeStamp;
         this.accountID = accountID;
+        this.toAccountID = toAccountID;
     }
     
     @Override
@@ -27,14 +30,13 @@ public class DepositLog implements Log {
             writer.write("Log ID: " + id + "\n");
             writer.write("Message: " + message + "\n");
             writer.write("Timestamp: " + timeStamp.toString() + "\n");
-            writer.write("Account ID: " + accountID + "\n");
+            writer.write("From Account ID: " + accountID + "\n");
+            writer.write("To Account ID: " + toAccountID + "\n");
             writer.write("----\n");
-            
         } 
         catch (IOException e) 
         {
             e.printStackTrace();
         }
     }
-    
 }
